MB-47092: Separate engine stats into high/low cardinality
Previously, all engine stats in the empty-key/"all" were exposed in the
low cardinality stat group as they are per-bucket.

As the number of KV metrics is quite high, the space used to retain all
KV metrics for a reasonable length of time is unacceptable.

As a step towards improving this, split engine stats into high/low
cardinality. This does not affect cbstats, but will move less critical
stats to the high cardinality endpoint, where they will be gathered
less frequently.

Stats which are now in the high cardinality endpoint:

* all config-derived metrics

* subset of engine stats:

    kv_ep_storage_age_seconds
    kv_ep_storage_age_highwat_seconds
    kv_ep_num_workers
    kv_ep_vbucket_del_max_walltime_seconds
    kv_ep_vbucket_del_avg_walltime_seconds
    kv_ep_num_access_scanner_runs
    kv_ep_num_access_scanner_skips
    kv_ep_access_scanner_last_runtime_seconds
    kv_ep_access_scanner_num_items
    kv_ep_defragmenter_num_visited
    kv_ep_defragmenter_num_moved
    kv_ep_defragmenter_sv_num_moved
    kv_ep_item_compressor_num_visited
    kv_ep_item_compressor_num_compressed
    kv_ep_cursor_dropping_lower_threshold_bytes
    kv_ep_cursors_dropped
    kv_ep_cursor_memory_freed_bytes
    kv_ep_mem_low_wat_percent_ratio
    kv_ep_mem_high_wat_percent_ratio
    kv_ep_mem_tracker_enabled
    kv_ep_bg_num_samples
    kv_ep_bg_min_wait_seconds
    kv_ep_bg_max_wait_seconds
    kv_ep_bg_wait_avg_seconds
    kv_ep_bg_min_load_seconds
    kv_ep_bg_max_load_seconds
    kv_ep_bg_load_avg_seconds

This change decreased the disk used by Prometheus on an idle node
with a single bucket after ~40h uptime by ~27% (18MB->13MB).

Change-Id: I1bc2dddb200ff6c01fcdf387a5f7f6529a55762e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156039
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>