MB-46729: Remove ep_tasks:uptime_s stats
The uptime_s used the ep_time functions initialized as part of
creation of ep-engine. Now that we want to be able to use the
executor _without_ creating instances of ep-engine we don't
want the executor to behave differently when we don't have
an instance of a ep bucket.

If it wasn't for the case that we also have a ton of unit
tests which use the executor in an isolated context (and
they may want to mock the timers in ep-engine) we could
have used the timer in the core.

Given that the _only_ use of the timer functions in the executor
is to get the current uptime of the server, we can just nuke
that (as it is already part of the stats returned from the
default stats request)

Change-Id: I141e926cf61ecd26359a085d94f7297a40b4e481
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155172
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>