MB-45026: Expelling doesn't remove the checkpoint_start item
A good amount of logic around checkpoint assumes the invariant that
the checkpoint_start item always exists in Checkpoint, same as the
empty-item.

When we implemented item-expel we broke that invariant. Expel may remove
the checkpoint_start item because it always assumes that it has to
remove all items that share the same seqno.
We still keep that logic at "expel", but not for the checkpoint_start
item that stays untouched now.

The change requires to preserve our cursor-registering logic by updating
the Checkpoint::getLowSeqno() logic accordingly to the new expel
behaviour.

Change-Id: Ic9dd5534b8888a416f93745e42b13ba1aaf3d324
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155204
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>