MB-46675: Generate timestamps for Prometheus metrics
Metric exporters may choose to include a timestamp for each exposed
sample; if omitted, Prometheus assigns a timestamp to each sample.

This patch exposes the time the metrics request was received by KV as
the timestamp for all returned samples.

This is in preparation for a later patch which will round the exposed
timestamp: exploration of Prometheus' on disk chunk file format has
found that rounding the exposed timestamps can significantly reduce
disk space requirements.

As ns_server requires the option to request non-timestamped metrics to
proxy to an external Prometheus instance, this patch introduces two new
metrics endpoints. KV previously exposed:

 /_prometheusMetrics
 /_prometheusMetricsHigh

and now additionally exposes:

 /_prometheusMetricsNoTS
 /_prometheusMetricsHighNoTS

Which are the variants of the existing endpoints which do not expose
timestamps.

It would be preferable to use a query parameter (e.g.,
?timestamps=false) but prometheus-cpp does not currently expose query
parameters to metric handlers.

Change-Id: I939ebea77281b97dff3617eca1fbb7088ab3946f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155944
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker