MB-46675: Round Prometheus metric timestamps to nearest second
Exploration of the Prometheus on disk chunk file format has found that
timestamps may constitute a large proportion of the stats disk usage.

Chunk file timestamps are encoded as delta-of-deltas (i.e., the
difference in scrape interval from the previous sample) in a variable
bit width field.

Inspecting chunk files from a 30 bucket idle node (see MB for logs)
found that a large proportion of DoDs were rather small:
-15ms <= x <= 16ms.

Because of the potential widths for the field on disk[1], such DoDs are
padded out to 14 bits + 2 bit prefix.

It has been established experimentally that rounding timestamps to the
nearest 100ms leads to a large proportion of these DoDs being exactly
zero - the small "jitter" is hidden. A zero DoD is encoded much more
efficiently on disk, as a single bit per sample. This constitutes a
significant reduction in disk usage.

In an entirely idle node with minimal load, there is little jitter and
so relatively little to gain, on the order of 10%. However, in more
heavily loaded nodes where Prometheus can no longer achieve as
precisely timed metric scrapes, the saving may be significant - the
absolute max would be ~16x less disk space required for timestamps,
though a real-world node is likely to see a more modest improvement.

The small loss of precision from rounding is considered acceptable;
given the scrape interval is normally 10s or greater, the impact on
rate calculations should be minimal.

The rounding does not lead to accumulated error - KV always rounds the
_current_ time, Prometheus then determines the timestamp delta and
delta-of-deltas using this value.

[1]:https://github.com/prometheus/prometheus/blob/release-2.26/tsdb/chunkenc/xor.go#L177

Change-Id: I0e2276404c3dfd22e004c813bba1ebcaaee374ec
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155945
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Restriction Checker