Refactor: remove the use of gsl::not_null<const void*> for AddStatFn
This patch refactors AddStatFn so that its third arg is a const void*
with the name ctx. As in our legacy uses of our code it was used a
context for AddStatFn as well as for cookies. Also by removing the
gsl::not_null<> we allow the user not to pass a context to the stats
function.

This patch also refactors mainly tests and our snapshot stats
functionality to use lambda functions that can bind to local variables
so that they may store data into a std::map<> for instance. This
increases data type consistency.

Change-Id: Ie046995dca08ef85eedd9e7192f63df62e6a47e4
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155294
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>