MB-43677: Remove NO_OPTIMIZE flags
When we introduce the precompiled header target GCC will complain about
a mismatch in flags between the precompiled header targets (which will
not have a NO_OPTIMIZE flag) and the targets that currently have set a
NO_OPTIMIZE flag. As this flag was included due to slow GTest compile
times and GTest will be going into the precompiled header we should
remove it, but in a separate patch to establish a correct baseline.

Before:

  Analyzing build trace from 'baseline.bin'...
**** Time summary:
Compilation (1066 times):
  Parsing (frontend):         3031.6 s
  Codegen & opts (backend):    800.4 s

After:

Analyzing build trace from 'noOptimize.bin'...
**** Time summary:
Compilation (1065 times):
  Parsing (frontend):         3093.3 s
  Codegen & opts (backend):    832.9 s

Change-Id: I088dc4ed5ae4de1b7d3a837a4999ad18bdc0d72e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155518
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>