MB-46827: Store reference to KVBucket in VBucket
In subsequent patches it will be useful to access the owning KVBucket
from the owned VBucket for checking the KVBucket memory state before
accepting or rejecting a mutation.

We could avoid this patch by checking the memory state at KVBucket
level, but VBucket add/set is where all the "store" operations flow and
where we already make the OOM checks, so I'm keeping the current
approach.

We already store a ref of KVBucket in EPBucket, so here I just move it
up in the hierarchy to VBucket.

Change-Id: I359b1c6995e2320b68f9ca696522fb418234ecf2
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155543
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>