MB-46900: Don't interpret BucketLogger prefix as fmt string
As part of "MB-46216: Check log format strings at compile-time
(ep-engine)" (f038ea0), the behaviour of BucketLogger when it has a
prefix which contained fmtlib formatting characters
changed. Previously this would throw an exception which spdlog would
catch, and print a log message along the lines of:

    "[*** LOG ERROR #0001 ***] [2021-06-15 19:06:07] [globalBucketLogger] {argument not found}"

However, with the changes in f038ea0 we perform the formatting of the
message outselves in bucket_logger.cc, and the thrown exception isn't
caught and can result in kv_engine terminating.

Fix by changing how we add the BucketLogger prefixes - don't append
them to the format string but instead explicilty format everything,
then pass the formatted log string down to spdlog.

Change-Id: Iae8212cd02e1ff9c28ae12921a10cedb7d3c6b0f
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155932
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>