MB-46827: Remove replication write-queue caps
- replication_throttle_cap_pcnt
- replication_throttle_queue_cap

Legacy and unused (disabled by default).

That was a way of implementing a replication-backoff machanism based on
a threshold on checkpoints.
The current (and more general) plan is to cap the mem-usage of
checkpoints, which is going to implement a backoff mechanism for both
master (frontend ops) and replica (DCP traffic).

Change-Id: Ic55087b9df01c14fd53dc7156be3a9b89a53327c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156006
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>