MB-42657: Make breakpad dumpCallback POSIX signal safe
On Linux replace usage of spdlog with calls to write to stderr as
spdlog is not signal safe.

This result in the loss of CRITICAL messages in the memcached.log
explaining the crash but retains the messages in the babysitter log.
The babysitter will timestamp the message and process exit. With the
loss of the CRITICAL messages on crash, the patch updates warmup to
write a single ERROR if an unclean shutdown was detected.

For Windows, the dumpCallback runs on a thread that is not the
signal handler and as such, is safe to use spdlog so remains unchanged.

Change-Id: Ic8f63e287606a62955bc8315f0c5cdd518542699
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155804
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>