MB-46988: Fix CLion project indexing with CB_PCH
When using precompiled headers and CLion; when it attempts to
preprocess each source file for indexing / analysis it fails with an
error about being unable to find <gtest/gtest.h>.

The issue here seems to be related to how CLion invokes the
preprocessor - it needs the include path for all headers inside
precompiled_headers.h; even though that has already been precompiled.

If we make the PCH target include paths public (i.e. so they are
inherited by consumers of the PCH target), then this addresses the
issue.

Change-Id: I7b86b140b53447e0ec22a242e275838c44e91c9d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156119
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>