MB-47056: Remove ERROR msg due to state of ep_force_shutdown
Remove the ERROR log message "Warmup detected a prior unclean
shutdown". As it was being logged, during warm-up of a new EP bucket.
This was due to the fact that new EP buckets don't have a stats.log and
thus, we wouldn't find a "ep_force_shutdown=false" key.

The message was originally added to help notify that memcached last
shutdown wasn't clean. But this will be covered by new code in before
this in an earlier phase of bucket initialization.

Change-Id: I781c13cf0c4ca6b4fe82192369208a52d1a95f9e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156335
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>