Allow breakpad test to run without installing
The breakpad test did not provide a root directory in the
configuration for memcached, so it would use the "compiled in"
value which is set to the install directory to look for
the error maps.

By providing a path to the source directory we may pick up
the error maps from the source directory instead.

Change-Id: I6661cc53bc37be97f0f494f0da207316edd65692
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156472
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>