MB-29928: Switch fragmentation scoring/weighting to use high-water mark
The justification for this change is based on current KV behaviour
and some input from a few 'random' real life log snapshots.

1) Current behaviour of KV's item-pager is to try and keep memory
below the high-water mark, at times of memory pressure it will attempt
to move memory to the low-water mark. The defragmenter cannot affect
mem_used, but it can affect the bucket's resident memory. Here we
allow the defragmenter's sleep to become most aggressive when
the resident memory is close/above the high-water mark.

2) It's not uncommon for users to have their mem_used between
the low and high water marks. This change means that the defragmenter
won't just be at it's most 'aggressive' for the users that are happily
between the two, only when they approach the high-water mark will
the scoring/weighting become higher.
Change-Id: Id942db14fe76ca6e18265fb66e379c18eecb6fab

Change-Id: I1f8d9098edf38dbd780b746a85da6a1177850482
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156643
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>