Add mcbp_fuzz_test and mcbp_network_fuzz_test
mcbp_fuzz_test is a small test program which use libFuzzer
(https://llvm.org/docs/LibFuzzer.html) to test the package
validation code.

The program is not built by default, and you need to explicitly
add -DCB_LIBFUZZER=ON when invoking CMake (Note: you should also
be using clang and have libfuzzer available)

Once built it may be run with:

    ./mcbp_fuzz_test -max_len=32

(There isn't much value of larger values as the packet header
is the first 32 bytes).

To make sure that we run as much as the server code as possible
the signature for Cookie::validate() was changed to allow bypass
of logging and exceptions for illegal packets.

mcbp_network_fuzz_test is similar to mcbp_fuzz_test, but it'll
forward the packets to "localhost:11210" where you can run
memcached with address sanitizer and have it check everything
in the full context (currently it don't try to auth and select
bucket so you need "default bucket" configured if you want it
to try to execute the packets, but the odds of actually get
a correct packet is probably really low ;))

Change-Id: I74da642ce6f9623feb4598549418285e3ac5f12e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156557
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>