MB-47183: Permit collection encoding of getAllVBSeqs at all times
Remove the "collection_enabled" check and allow the collection
encoding at all times. The check was intended to hide the extra
encoding during the 6.5 DP, but can now be removed. If collections
are not enabled (e.g. during upgrade) it is ok, the requested
collection will be checked against what we know exists.

Change-Id: I96d6742defa81d6a7d59d2043790881618c85c68
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157153
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>