MB-47109: Fix cbstats from crashing due to 'ep_tasks:uptime_s'
Stat 'ep_tasks:uptime_s' was removed in commit
de9b8c13348dce593623b830238e6d782b4b7c17. This causes cbstats to crash
as it tried to access the stat in the data it received for 'tasks'.
To fix this crash, remove the accessing of 'ep_tasks:uptime_s' and its
printing as part of the tasks data. This will be compatible with
mad-hatter/CC server versions. But will not print the up time.

Change-Id: Ic98ae99d2b93903ad9bd8f2d04b7b0f3c227cd49
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156704
Tested-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>