MB-47216: Don't validate all packets
As demonstrated in MB-47151 an unauthenticated client may exploit
bugs in packet validators and potentially crash the system.

This patch change the logic so that we'll return EACCESS if the
connection isn't authenticated (and default bucket isn't available)
and the command isn't whitelisted.

The following commands will always be validated (and potentially
excecuted):

   * Hello
   * Sasl List Mech
   * Sasl Auth
   * Sasl Step

Change-Id: Ifd40b177dd1312c11e9b2f2f707f67504df6e2b7
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/156867
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>