MB-47143: Add Rate Limited error codes
The following error codes was added to the system:

    RateLimitedNetworkIngress = 0x30,
    RateLimitedNetworkEgress = 0x31,
    RateLimitedMaxConnections = 0x32,
    RateLimitedMaxCommands = 0x33,

A new attribute is added to the error map (named "rate-limit")
to allow clients to "automatically" recognize new rate limits
when they are added to the system and perform "magic" related
to them (like trigger a callback for the user)

Given that not all clients support new attributes, the version
of the clustermap was bumped to two and in the spec for version
two clients _MUST_ ignore unknown attributes. The rest of the
layout is identical to version 1.

To avoid having to maintain two different formats on the server
we generate version 1 out of version 2 during startup. Clients
requesting version 2 must accept the server returning version 1.

Change-Id: I1fafa86dd8b85fc2f5474321637aaaf237ce1069
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157226
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>