MB-46827: Track bucket checkpoint mem-usage in EPStats::coreLocal
For limiting the checkpoint mem-usage we need to checkout the
checkpoint memory status on frontend threads. So, any expensive way of
getting that information must be avoided.

We add a new estimate of the checkpoint mem-usage in EPStats::coreLocal
and use that. Writes are being fast and reads requires a few iterations
on threads-local for computing the aggregate.

In the estimate we account for the memory of items queued in the
checkpoint and for the allocation in the key-index.

Change-Id: I85ebe8d5c5c5ef2cf2c2989433be2c5b92ee0e1b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157352
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>