MB-47267: Make backfill during warmup a PauseResume task
MB-47267 demonstrated that during warmup of a node in a cluster which
has multiple buckets of different sizes, smaller buckets could become
stuck behind larger buckets causing their warmup time to be much longer
than it need be. This can causes problems during delta node recovery as
NS_Server starts requesting vbucket stats to check vbucket state and
if this is not responded to promptly this will cause the rebalance to
timeout. Which may happen as EP-Engine does not service any vbucket
stats requests till we've completed 'PopulateVBucketMap' state of
warmup, which requires  one + five x the number of shards reader tasks
to have been run and completed. Though these tasks are relatively quick
to run, they can become stuck behind other bucket's warmup backfilling
tasks which can take upwards of minutes to run depending on the size of
the bucket. As they share the same reader thread task pool.

To fix this, this patch modifies the backfilling functionality during
warmup so that backfill may now be paused after a time deadline has
been met and store the seqno we've reached so that we can start
backfilling again when the tasks is rescheduled. Which in turn allows
the faster initial phases of warmup to have a chance of being run
between each run of backfill.  This patch unifies the functionality
of the backfilling stages of warmup so that the same code is used for
'KeyDump', 'LoadingKVPairs' and 'LoadingData' stages of warmup.

This patch also adds a test to ensure that, if we create a larger
bucket, and start warmup it up before a smaller one. With only one
reader thread that the smaller bucket is able to warmup before that
the larger one has finished warming up.

A ep_perfsuite test has been added to test warmup latency to help with
tuning of the maxDuration time for that a backfill can take before it's
paused.

Change-Id: If983145100e8f871d2c49caa5da9f744c6134b4b
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/158362
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>