MB-46474: Remove DCP stats from snapshot stats
The DCP stats group contains details about every DCP connection and
stream currrently open. This can easily reach 100,000s of elemnts for
even modest workloads), and consequently the amount of memory required
to serialise them can be of the order of 100MB+.

This can be a significant jump in mem_used, particulary for buckets
with a small memory quota; which can result in mem_used exceeding the
high watermark and many items being evicted.

Given I can't recall an occasion when the dcp content in stats.json
has actually been used to debug anything; and we have essentially
equivilent information in stats.log recorded by cbcollect_info; simply
nix the DCP stats from those which are snapshotted and persisted.

Change-Id: Idc3adac922daed501b6e44067fb738bb95cb33e8
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157437
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>