MB-46827: Introduce KVBucket::hasCapacityInCheckpoints()
The function verifies whether checkpoints memory conditions allow a new
item to be enqueued for persistence and replication.

With this patch we start enforcing the new checkpoint quota for ADD/SET
operations.
A follow-up patch will cover DEL/EXP.

Note that currently the checkpoint_memory_ratio is set to 1.0 in
configuration, so the memory behaviour of the system stays logically
unchanged.

Change-Id: I29368cdee147fe6421398a75aff25f47e00725ca
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/155544
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>