MB-46827: Remove cursor_dropping_<lower/upper>_mark
With bounded checkpoint mem-usage we are going to trigger checkpoint
memory recovery only when the checkpoint mem-usage is high.

The params that I'm removing here are used for triggering checkpoint
memory recovery also when the overall mem-usage of the system (not the
checkpoints one specifically) is high.
Which is a behaviour that we want to remove: if some other components
shows up high mem-usage then we should release memory from there, not
from checkpoints.

Here I also remove some ep_testuite tests that specifically verify the
behaviour of the params that I'm removing. A new version of those tests
will be re-introduced (possibly as ep-unit-tests) at the right time in
following patches to verify the new triggers for checkpoint
mem-recovery.

Change-Id: I9381b9d74b0ffa48ec8f006a86a6a52cbb46d981
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157520
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>