MB-33784: Add StorageProperties::AutomaticDeduplication
We can avoid optimizing writes before every flush if the underlying
kvstore (e.g. Rocks) will dedupe for us. Magma used to be able to
do this too, but the functionality was removed in recent
optimizations around group commit. The magma team will investigate
this in the future though with the intent to enable it.

Change-Id: I01dbdd73a112f6ab7ba90517517798729ebe8d7d
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157581
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>