MB-33784: Empty collection same flush optimization checks itemInBatch
We currently detect when we drop an empty collection so that we can
avoid triggering a purge (compaction) if there is meaningful work to
be done. An empty collection may exist in one of two scenarios:

1) The collection previously existed on disk and was dropped in some
   flush batch

2a) The collection was created and dropped in the same flush batch

Currently in scenario 2(a) we will persist only a drop collection system
event and the flush code will check the high seqno when processing
the collection data to determine if it should trigger a purge or not.
A highSeqno of 0 means nothing was written to the collection and no
purge is required.

With the changes being made under this MB to only perform flusher
dedupe for the storage backends that require it, a new scenario,
scenario 2(b) is introduced in which a collection is created and
dropped in the same flush batch, but the store persists both system
events. In this scenario the highSeqno check against 0 does not work
as it is set for the create system event.

Switch the check for scenario 2 from highSeqno to a new bool tracked
for each flush (itemInBatch).

Change-Id: I74ff8bb9d3800eaf92851c2c525f5e60039cb065
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157580
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>