MB-47308: Use MemoryDomain::Secondary for libmagma
Create a proxy class that MagmaKVStore will use for accessing the real
Magma. The proxy class will ensure that the Secondary domain is
selected as we call into libmagma.

In this first patch not all function calls are executing in the
secondary domain as further work is required to ensure any heap memory
returned from the library is freed in the correct domain.

Change-Id: I42c955317670f822e1be080102053c97973fde95
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157582
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>