MB-33784: Don't dedupe in flusher if not required
Some backend KVStores will perform dedupe for us so doing the work
in the flusher is unnecessary.

Many tests assert the sizes of flush batches (as reported by the
flusher and not the KVStore) and these numbers will now have changed
in any tests that relied in some way on flusher dedupe. These tests
now check the expected behaviour and supply the correct expectation
on the flush batch size.

Change-Id: I71676a82256cb6bd4e4911df0132849e93b4f900
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157651
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>