MB-42652: Magma: Only remove dropped collections at end of compaction
Currently during compaction (which is not inter-locked with flushing)
we delete the dropped collections doc after purging any collections.
If we flush a drop of a collection during this compaction then we
may effectively miss updating the dropped collections doc by deleting
it after the compaction finishes. This means that a subsequent
compaction would not purge the dropped collection and data would
remain on disk and the collection items would keep contributing
towards the item count.

Correct this by only removing the collections that we drop from the
dropped collections stat doc at the end of the compaction, whilst
holding the vBucket write lock to prevent races with the flusher.

Change-Id: I815cf63135a388b224c7162868868f23498e7182
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157705
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>