MB-47484: Don't use Executorpool before created
The settings callbacks set up in settings_init will be
used during the initial load of the configuration file and parts
of the system may not be initialized at the time.

Refactor the code so that we delay the installation of the
callbacks until the pool is created (note: It still cannot
be called from a users context as we haven't started dispatching
client connections yet).

Added the parameters to the testapp config to verify that the
bug is fixed

Change-Id: I720a45c9cb33bf5b65d590bc6f369e9a0fab148c
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157835
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>