MB-47386: Remove 'openCkptCreated' out-arg from ckpt-removal function
Used in production code only in legacy code that was responsible for
notifying TAP connections for sending checkpoint start/end items
(http://review.couchbase.org/c/ep-engine/+/4970/).

Rarely used in test code. In the few cases where we need it, we can just
check if the open-checkpoint-id has changed in the test-code rather than
in the production code.

Note that this patch is also in preparation for splitting the existing
CM::removeClosedUnrefCheckpoints function for enhanced testing.

Change-Id: I1f21e66a0ef594bdac34f6e0fd08d9562504aaa3
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/157853
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>